Done with Bridge Work Summary Tab
