+++ BridgeCareApp/BridgeCare/BridgeCare.csproj
@@ -238,6 +238,10 @@
     <Compile Include="Services\DetailedReport\Target.cs" />
     <Compile Include="Services\Network.cs" />
     <Compile Include="Services\Simulations.cs" />
+    <Compile Include="Services\SummaryReport\BridgeRateDeckAreaWorkSummary.cs" />
+    <Compile Include="Services\SummaryReport\BridgesCulvertsWorkSummary.cs" />
+    <Compile Include="Services\SummaryReport\BridgeWorkSummaryCommon.cs" />
+    <Compile Include="Services\SummaryReport\CostBudgetsWorkSummary.cs" />
     <Compile Include="Services\SummaryReport\ExcelHelper.cs" />
     <Compile Include="Services\SummaryReport\SummaryReportBridgeData.cs" />
     <Compile Include="Services\SummaryReport\BridgeWorkSummary.cs" />
