+++ BridgeCareApp/BridgeCare/Models/SummaryReport/YearsData.cs
@@ -10,7 +10,7 @@
 
         public string Sub { get; set; }
 
-        public string Culv { get; set; }
+        public string Culv { get; set; }        
 
         public string DeckD { get; set; }
